✅ addTweet Func New js file update
